Refactor secrets management for MongoDB connection
The commit makes some changes to the way we handle secrets for MongoDB connections in AWS DocumentDB. Instead of storing the complete MongoDB URL in the Secrets Manager, the commit breaks the URL into pieces and stores them as separate secrets (username, password, host, port, replica set). It also provides the respective outputs for these values.